lang: move accounts into their own namespace
