+++ lang/src/accounts/mod.rs
@@ -1,0 +1,24 @@
+pub mod account;
+pub mod account_info;
+pub mod boxed;
+#[doc(hidden)]
+#[allow(deprecated)]
+pub mod cpi_account;
+#[doc(hidden)]
+#[allow(deprecated)]
+pub mod cpi_state;
+#[doc(hidden)]
+#[allow(deprecated)]
+pub mod loader;
+pub mod loader_account;
+pub mod program;
+#[doc(hidden)]
+#[allow(deprecated)]
+pub mod program_account;
+pub mod signer;
+#[doc(hidden)]
+#[allow(deprecated)]
+pub mod state;
+pub mod system_account;
+pub mod sysvar;
+pub mod unchecked_account;
