+++ lang/src/accounts/state.rs
@@ -1,10 +1,10 @@
+#[allow(deprecated)]
+use crate::accounts::cpi_account::CpiAccount;
 use crate::error::ErrorCode;
 use crate::{
     AccountDeserialize, AccountSerialize, Accounts, AccountsExit, Key, ToAccountInfo,
     ToAccountInfos, ToAccountMetas,
 };
-#[allow(deprecated)]
-use crate::{CpiStateContext, ProgramState};
 use solana_program::account_info::AccountInfo;
 use solana_program::entrypoint::ProgramResult;
 use solana_program::instruction::AccountMeta;
@@ -12,11 +12,12 @@ use solana_program::program_error::ProgramError;
 use solana_program::pubkey::Pubkey;
 use std::ops::{Deref, DerefMut};
 
-/// Boxed container for the program state singleton, used when the state
-/// is for a program not currently executing.
+pub const PROGRAM_STATE_SEED: &str = "unversioned";
+
+/// Boxed container for the program state singleton.
 #[derive(Clone)]
 #[deprecated]
-pub struct CpiState<'info, T: AccountSerialize + AccountDeserialize + Clone> {
+pub struct ProgramState<'info, T: AccountSerialize + AccountDeserialize + Clone> {
     inner: Box<Inner<'info, T>>,
 }
 
@@ -27,49 +28,51 @@ struct Inner<'info, T: AccountSerialize + AccountDeserialize + Clone> {
 }
 
 #[allow(deprecated)]
-impl<'info, T: AccountSerialize + AccountDeserialize + Clone> CpiState<'info, T> {
-    pub fn new(i: AccountInfo<'info>, account: T) -> CpiState<'info, T> {
+
+impl<'a, T: AccountSerialize + AccountDeserialize + Clone> ProgramState<'a, T> {
+    fn new(info: AccountInfo<'a>, account: T) -> ProgramState<'a, T> {
         Self {
-            inner: Box::new(Inner { info: i, account }),
+            inner: Box::new(Inner { info, account }),
         }
     }
 
-    /// Deserializes the given `info` into a `CpiState`.
+    /// Deserializes the given `info` into a `ProgramState`.
     #[inline(never)]
-    pub fn try_from(info: &AccountInfo<'info>) -> Result<CpiState<'info, T>, ProgramError> {
+    pub fn try_from(
+        program_id: &Pubkey,
+        info: &AccountInfo<'a>,
+    ) -> Result<ProgramState<'a, T>, ProgramError> {
+        if info.owner != program_id {
+            return Err(ErrorCode::AccountNotProgramOwned.into());
+        }
+        if info.key != &Self::address(program_id) {
+            solana_program::msg!("Invalid state address");
+            return Err(ErrorCode::StateInvalidAddress.into());
+        }
         let mut data: &[u8] = &info.try_borrow_data()?;
-        Ok(CpiState::new(info.clone(), T::try_deserialize(&mut data)?))
+        Ok(ProgramState::new(
+            info.clone(),
+            T::try_deserialize(&mut data)?,
+        ))
     }
 
-    fn seed() -> &'static str {
-        ProgramState::<T>::seed()
+    pub fn seed() -> &'static str {
+        PROGRAM_STATE_SEED
     }
 
     pub fn address(program_id: &Pubkey) -> Pubkey {
-        let (base, _nonce) = Pubkey::find_program_address(&[], program_id);
-        let seed = Self::seed();
-        let owner = program_id;
-        Pubkey::create_with_seed(&base, seed, owner).unwrap()
-    }
-
-    /// Convenience api for creating a `CpiStateContext`.
-    pub fn context<'a, 'b, 'c, A: Accounts<'info>>(
-        &self,
-        program: AccountInfo<'info>,
-        accounts: A,
-    ) -> CpiStateContext<'a, 'b, 'c, 'info, A> {
-        CpiStateContext::new(program, self.inner.info.clone(), accounts)
+        address(program_id)
     }
 }
 
 #[allow(deprecated)]
-impl<'info, T> Accounts<'info> for CpiState<'info, T>
+impl<'info, T> Accounts<'info> for ProgramState<'info, T>
 where
     T: AccountSerialize + AccountDeserialize + Clone,
 {
     #[inline(never)]
     fn try_accounts(
-        _program_id: &Pubkey,
+        program_id: &Pubkey,
         accounts: &mut &[AccountInfo<'info>],
         _ix_data: &[u8],
     ) -> Result<Self, ProgramError> {
@@ -78,17 +81,13 @@ where
         }
         let account = &accounts[0];
         *accounts = &accounts[1..];
-
-        // No owner or address check is done here. One must use the
-        // #[account(state = <account-name>)] constraint.
-
-        CpiState::try_from(account)
+        ProgramState::try_from(program_id, account)
     }
 }
 
 #[allow(deprecated)]
 impl<'info, T: AccountSerialize + AccountDeserialize + Clone> ToAccountMetas
-    for CpiState<'info, T>
+    for ProgramState<'info, T>
 {
     fn to_account_metas(&self, is_signer: Option<bool>) -> Vec<AccountMeta> {
         let is_signer = is_signer.unwrap_or(self.inner.info.is_signer);
@@ -102,7 +101,7 @@ impl<'info, T: AccountSerialize + AccountDeserialize + Clone> ToAccountMetas
 
 #[allow(deprecated)]
 impl<'info, T: AccountSerialize + AccountDeserialize + Clone> ToAccountInfos<'info>
-    for CpiState<'info, T>
+    for ProgramState<'info, T>
 {
     fn to_account_infos(&self) -> Vec<AccountInfo<'info>> {
         vec![self.inner.info.clone()]
@@ -111,7 +110,7 @@ impl<'info, T: AccountSerialize + AccountDeserialize + Clone> ToAccountInfos<'in
 
 #[allow(deprecated)]
 impl<'info, T: AccountSerialize + AccountDeserialize + Clone> ToAccountInfo<'info>
-    for CpiState<'info, T>
+    for ProgramState<'info, T>
 {
     fn to_account_info(&self) -> AccountInfo<'info> {
         self.inner.info.clone()
@@ -120,7 +119,7 @@ impl<'info, T: AccountSerialize + AccountDeserialize + Clone> ToAccountInfo<'inf
 
 #[allow(deprecated)]
 impl<'info, T: AccountSerialize + AccountDeserialize + Clone> AsRef<AccountInfo<'info>>
-    for CpiState<'info, T>
+    for ProgramState<'info, T>
 {
     fn as_ref(&self) -> &AccountInfo<'info> {
         &self.inner.info
@@ -128,7 +127,7 @@ impl<'info, T: AccountSerialize + AccountDeserialize + Clone> AsRef<AccountInfo<
 }
 
 #[allow(deprecated)]
-impl<'info, T: AccountSerialize + AccountDeserialize + Clone> Deref for CpiState<'info, T> {
+impl<'a, T: AccountSerialize + AccountDeserialize + Clone> Deref for ProgramState<'a, T> {
     type Target = T;
 
     fn deref(&self) -> &Self::Target {
@@ -137,24 +136,45 @@ impl<'info, T: AccountSerialize + AccountDeserialize + Clone> Deref for CpiState
 }
 
 #[allow(deprecated)]
-impl<'info, T: AccountSerialize + AccountDeserialize + Clone> DerefMut for CpiState<'info, T> {
+impl<'a, T: AccountSerialize + AccountDeserialize + Clone> DerefMut for ProgramState<'a, T> {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut DerefMut::deref_mut(&mut self.inner).account
     }
 }
 
 #[allow(deprecated)]
-impl<'info, T: AccountSerialize + AccountDeserialize + Clone> AccountsExit<'info>
-    for CpiState<'info, T>
+impl<'info, T> From<CpiAccount<'info, T>> for ProgramState<'info, T>
+where
+    T: AccountSerialize + AccountDeserialize + Clone,
 {
-    fn exit(&self, _program_id: &Pubkey) -> ProgramResult {
-        // no-op
-        Ok(())
+    fn from(a: CpiAccount<'info, T>) -> Self {
+        Self::new(a.to_account_info(), Deref::deref(&a).clone())
     }
 }
 
 #[allow(deprecated)]
-impl<'info, T: AccountSerialize + AccountDeserialize + Clone> Key for CpiState<'info, T> {
+impl<'info, T: AccountSerialize + AccountDeserialize + Clone> AccountsExit<'info>
+    for ProgramState<'info, T>
+{
+    fn exit(&self, _program_id: &Pubkey) -> ProgramResult {
+        let info = self.to_account_info();
+        let mut data = info.try_borrow_mut_data()?;
+        let dst: &mut [u8] = &mut data;
+        let mut cursor = std::io::Cursor::new(dst);
+        self.inner.account.try_serialize(&mut cursor)?;
+        Ok(())
+    }
+}
+
+pub fn address(program_id: &Pubkey) -> Pubkey {
+    let (base, _nonce) = Pubkey::find_program_address(&[], program_id);
+    let seed = PROGRAM_STATE_SEED;
+    let owner = program_id;
+    Pubkey::create_with_seed(&base, seed, owner).unwrap()
+}
+
+#[allow(deprecated)]
+impl<'info, T: AccountSerialize + AccountDeserialize + Clone> Key for ProgramState<'info, T> {
     fn key(&self) -> Pubkey {
         *self.inner.info.key
     }
